SOLR-12325 Refactored uniqueBlock agg classes
